CDK/requirements.txt
aws-cdk-lib>=2.147.1
constructs>=10.0.0,<11.0.0
